Moved some builder code into their own modules
