+++ pipelex/builder/builder.py
@@ -1,10 +1,9 @@
-from typing import TYPE_CHECKING, Annotated, cast
+from typing import TYPE_CHECKING, cast
 
-from pydantic import ConfigDict, Field, ValidationError, field_validator
+from pydantic import ValidationError
 
 from pipelex.builder.builder_errors import (
     ConceptDefinitionErrorData,
-    ConceptFailure,
     ConceptSpecError,
     DomainFailure,
     PipeBuilderError,
@@ -16,7 +15,9 @@ from pipelex.builder.builder_errors import (
     StaticValidationErrorData,
     ValidateDryRunError,
 )
+from pipelex.builder.bundle_spec import PipelexBundleSpec
 from pipelex.builder.concept.concept_spec import ConceptSpec
+from pipelex.builder.domain_spec import DomainSpec
 from pipelex.builder.pipe.pipe_batch_spec import PipeBatchSpec
 from pipelex.builder.pipe.pipe_compose_spec import PipeComposeSpec
 from pipelex.builder.pipe.pipe_condition_spec import PipeConditionSpec
@@ -27,13 +28,9 @@ from pipelex.builder.pipe.pipe_llm_spec import PipeLLMSpec
 from pipelex.builder.pipe.pipe_parallel_spec import PipeParallelSpec
 from pipelex.builder.pipe.pipe_sequence_spec import PipeSequenceSpec
 from pipelex.builder.pipe.pipe_signature import PipeSpec
-from pipelex.core.bundles.pipe_sorter import sort_pipes_by_dependencies
-from pipelex.core.bundles.pipelex_bundle_blueprint import PipeBlueprintUnion, PipelexBundleBlueprint
-from pipelex.core.concepts.concept_blueprint import ConceptBlueprint
-from pipelex.core.domains.domain_blueprint import DomainBlueprint
+from pipelex.builder.pipe.pipe_spec_union import PipeSpecUnion
 from pipelex.core.memory.working_memory import WorkingMemory
 from pipelex.core.stuffs.list_content import ListContent
-from pipelex.core.stuffs.structured_content import StructuredContent
 from pipelex.exceptions import (
     ConceptLoadingError,
     DomainLoadingError,
@@ -47,117 +44,6 @@ from pipelex.tools.typing.pydantic_utils import format_pydantic_validation_error
 
 if TYPE_CHECKING:
     from pipelex.core.stuffs.list_content import ListContent
-
-
-class DomainInformation(StructuredContent):
-    domain: str = Field(description="Name of the domain of the knowledge work.")
-    description: str = Field(description="Definition of the domain of the knowledge work.")
-
-
-PipeSpecUnion = Annotated[
-    PipeFuncSpec
-    | PipeImgGenSpec
-    | PipeComposeSpec
-    | PipeLLMSpec
-    | PipeExtractSpec
-    | PipeBatchSpec
-    | PipeConditionSpec
-    | PipeParallelSpec
-    | PipeSequenceSpec,
-    Field(discriminator="type"),
-]
-
-
-class PipelexBundleSpec(StructuredContent):
-    """Complete spec of a Pipelex bundle TOML definition.
-
-    Represents the top-level structure of a Pipelex bundle, which defines a domain
-    with its concepts, pipes, and configuration. Bundles are the primary unit of
-    organization for Pipelex workflows, loaded from TOML files.
-
-    Attributes:
-        domain: The domain identifier for this bundle in snake_case format.
-               Serves as the namespace for all concepts and pipes within.
-        description: Natural language description of the pipeline's purpose and functionality.
-        system_prompt: Default system prompt applied to all LLM pipes in the bundle
-                      unless overridden at the pipe level.
-        concept: Dictionary of concept definitions used in this domain. Keys are concept
-                codes in PascalCase format, values are ConceptBlueprint instances or
-                string references to existing concepts.
-        pipe: Dictionary of pipe definitions for data transformation. Keys are pipe
-             codes in snake_case format, values are specific pipe spec types
-             (PipeLLM, PipeImgGen, PipeSequence, etc.).
-
-    Validation Rules:
-        1. Domain must be in valid snake_case format.
-        2. Concept keys must be in PascalCase format.
-        3. Pipe keys must be in snake_case format.
-        4. Extra fields are forbidden (strict mode).
-        5. Pipe types must match their blueprint discriminator.
-
-    """
-
-    model_config = ConfigDict(extra="forbid")
-
-    domain: str
-    description: str | None = None
-    system_prompt: str | None = None
-
-    concept: dict[str, ConceptSpec | str] | None = Field(default_factory=dict)
-
-    pipe: dict[str, PipeSpecUnion] | None = Field(default_factory=dict)
-
-    @field_validator("domain", mode="before")
-    @classmethod
-    def validate_domain_syntax(cls, domain: str) -> str:
-        DomainBlueprint.validate_domain_code(code=domain)
-        return domain
-
-    def to_blueprint(self) -> PipelexBundleBlueprint:
-        concept: dict[str, ConceptBlueprint | str] | None = None
-
-        if self.concept:
-            concept = {}
-            for concept_code, concept_spec_or_name in self.concept.items():
-                if isinstance(concept_spec_or_name, ConceptSpec):
-                    try:
-                        concept[concept_code] = concept_spec_or_name.to_blueprint()
-                    except ValidationError as exc:
-                        msg = f"Failed to create concept blueprint from spec for concept code {concept_code}: {format_pydantic_validation_error(exc)}"
-                        concept_failure = ConceptFailure(concept_spec=concept_spec_or_name, error_message=msg)
-                        raise ConceptSpecError(message=msg, concept_failure=concept_failure) from exc
-                else:
-                    concept[concept_code] = ConceptBlueprint(description=concept_code, structure=concept_spec_or_name)
-
-        pipe: dict[str, PipeBlueprintUnion] | None = None
-        if self.pipe:
-            # First, convert all specs to blueprints
-            pipe_blueprints: dict[str, PipeBlueprintUnion] = {}
-            for pipe_code, pipe_spec in self.pipe.items():
-                try:
-                    pipe_blueprints[pipe_code] = pipe_spec.to_blueprint()
-                except ValidationError as exc:
-                    msg = f"Failed to create pipe blueprint from spec for pipe code {pipe_code}: {format_pydantic_validation_error(exc)}"
-                    pipe_failure = PipeFailure(pipe_spec=pipe_spec, error_message=msg)
-                    raise PipeSpecError(message=msg, pipe_failure=pipe_failure) from exc
-
-            # Then, sort blueprints by dependencies
-            try:
-                sorted_pipe_items = sort_pipes_by_dependencies(pipe_blueprints)
-            except Exception as exc:
-                msg = f"Failed to sort pipes by dependencies: {exc}"
-                raise PipeBuilderError(msg) from exc
-
-            # Finally, create the ordered dict
-            pipe = dict(sorted_pipe_items)
-
-        return PipelexBundleBlueprint(
-            domain=self.domain,
-            description=self.description,
-            system_prompt=self.system_prompt,
-            pipe=pipe,
-            concept=concept,
-        )
 
 
 # # TODO: Put this in a factory. Investigate why it is necessary.
@@ -203,7 +89,7 @@ async def assemble_pipelex_bundle_spec(working_memory: WorkingMemory) -> Pipelex
     )
 
     pipe_specs: list[PipeSpecUnion] = cast("ListContent[PipeSpecUnion]", working_memory.get_stuff(name="pipe_specs").content).items
-    domain_information = working_memory.get_stuff_as(name="domain_information", content_type=DomainInformation)
+    domain_information = working_memory.get_stuff_as(name="domain_information", content_type=DomainSpec)
 
     # Properly validate and reconstruct concept specs to ensure proper Pydantic validation
     validated_concepts: dict[str, ConceptSpec | str] = {}
@@ -220,6 +106,8 @@ async def assemble_pipelex_bundle_spec(working_memory: WorkingMemory) -> Pipelex
     return PipelexBundleSpec(
         domain=domain_information.domain,
         description=domain_information.description,
+        system_prompt=domain_information.system_prompt,
+        main_pipe=domain_information.main_pipe,
         concept=validated_concepts,
         pipe={pipe_spec.pipe_code: _convert_pipe_spec(pipe_spec) for pipe_spec in pipe_specs},
     )
