+++ pipelex/core/bundles/pipelex_bundle_blueprint.py
@@ -35,6 +35,7 @@ class PipelexBundleBlueprint(BaseModel):
     domain: str
     description: str | None = None
     system_prompt: str | None = None
+    main_pipe: str | None = None
 
     concept: dict[str, ConceptBlueprint | str] | None = Field(default_factory=dict)
